Depend on gson and commons-lang3 API plugins
The transitive dependencies included through the ironmq jar file are out
of date and won't be used in Jenkins installations where the API plugins
are already installed.  That will be the case in most Jenkins
controllers, so let's make this plugin use the API plugins rather than
packaging the commons-lang3 and gson jar files as transitive
dependencies.

The commons-lang3 plugin is installed on 65% of all Jenkins controllers
and is delivering the most recent release of commons-lang3.

The gson plugin is a new enough plugin that we don't have installation
count data yet, but it is expected to be widely installed as well,
since it is used in the Trilead API plugin that is installed on 93%
of Jenkins controllers.